py/pyproject.toml
[tool.poetry]
name = "recipamatic"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "recipamatic", from = "src"}]

[tool.poetry.dependencies]
python = "~3.11"
loguru = "^0.7.3"
rich = "^13.9.4"
tqdm = "^4.67.1"
ipykernel = "^6.29.5"
ipywidgets = "^8.1.5"
nbformat = "^5.10.4"
langchain = "^0.3.12"
langchain-community = "^0.3.12"
langchain-openai = "^0.2.12"
instaloader = "^4.14"
jsons = "^1.6.3"
llvmlite = "^0.43.0"
openai-whisper = "^20240930"
pydub = "^0.25.1"
py3langid = "<0.3"
fastapi = {extras = ["standard"], version = "^0.115.6"}
python-jose = {extras = ["cryptography"], version = "^3.4.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
pydantic-settings = "^2.8.1"
python-multipart = "^0.0.20"
authlib = "^1.5.2"

[tool.poetry.group.test.dependencies]
pytest = "^8.3.3"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"
force_sort_within_sections = true
multi_line_output = 3

